requirements.txt
beautifulsoup4
python-dotenv
python-telegram-bot
requests